requirements.txt
flask
pyzmq
requests
pillow
